Completes the weekly commit count chart
This is a stacked column chart, with the x-axis the week number, and the
y-axis number of commits in that week. The 2 series are number of commits
from the repo owner, and the number of commits from non-owners.

Some additional work on the visualisation and the values on the x-axis
will be required.