Removes the ShinobiCharts binary and replaces with Xamarin Component
This uses the trial so we've also got a ShinobiLicenseKeyProvider, which
reads the license keys from a JSON file. The old github-only file has
been repurposed to include the Shinobi details as well.

NOTE: This commit doesn't currently work, because we have code which
requires v2.3 of ShinobiCharts, but the version in the Xamarin component
store is currently 2.2.1. Once this has been updated I'll re-check this
commit and at that stage we can merge it back.